feat: componentização do código
