+++ src/assets/index.js
@@ -37,3 +37,4 @@ export { ReactComponent as AddDarkIcon } from './icons/add-dark-icon.svg';
 export { ReactComponent as CartButtonIcon } from './icons/cart-button-icon.svg';
 export { ReactComponent as DeliveryDarkIcon } from './icons/delivery-dark-icon.svg';
 export { ReactComponent as PeopleLightIcon } from './icons/people-light-icon.svg';
+export { ReactComponent as CloseIcon } from './icons/close-icon.svg';
