+++ src/pages/register/index.js
@@ -1,18 +1,11 @@
-import { useState, useRef } from 'react';
 import { RegisterContainer, Form, InputContainer } from './styles';
-import {
-  UserIcon,
-  CellIcon,
-  CadIcon,
-  EyeCloseIcon,
-  CadastroTextIcon,
-} from '../../assets/index';
+import { CadastroTextIcon } from '../../assets/index';
 import { ActionButton } from '../../styles/ComponentsStyles';
 import BreadCrumbs from '../../components/breadCrumbs';
+import Input from '../../components/form/input';
 
 import axios from '../../services/axios';
 import { useNavigate } from 'react-router-dom';
-import { Input } from '../../styles/ComponentsStyles';
 import { toast } from 'react-toastify';
 
 import { useForm } from 'react-hook-form';
@@ -20,16 +13,7 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { userSchema } from '../../validations/user-admin/userSchema';
 
 function Register() {
-  const [isVisible, setIsVisible] = useState(false);
-  const [confirmIsVisible, setConfirmIsVisible] = useState(false);
   const navigate = useNavigate();
-
-  const refs = {
-    name: useRef(null),
-    phone: useRef(null),
-    password: useRef(null),
-    confirmPassword: useRef(null),
-  };
 
   const {
     register,
@@ -39,11 +23,11 @@ function Register() {
     resolver: yupResolver(userSchema({ role: 'client', mode: 'register' })),
   });
 
-  async function onSubmit({ name, indentifier, password }) {
+  async function onSubmit({ name, phone, password }) {
     try {
       await axios.post('/clients', {
         name,
-        phone: indentifier,
+        phone,
         password,
       });
       toast.success('Conta criada com sucesso!');
@@ -73,49 +57,33 @@ function Register() {
           <CadastroTextIcon />
 
           <InputContainer>
-            <Input onClick={() => refs.name.current?.focus()}>
-              <UserIcon />
-              <input
-                ref={refs.name}
-                type="text"
-                placeholder="Nome"
-                {...register('name')}
-              />
-            </Input>
+            <Input
+              typeField="name"
+              placeholder="Nome"
+              register={register}
+              field="name"
+            />
 
-            <Input onClick={() => refs.phone.current?.focus()}>
-              <CellIcon />
-              <input
-                ref={refs.phone}
-                type="text"
-                placeholder="Telefone"
-                {...register('indentifier')}
-              />
-            </Input>
+            <Input
+              typeField="phone"
+              placeholder="Telefone"
+              register={register}
+              field="phone"
+            />
 
-            <Input onClick={() => refs.password.current?.focus()}>
-              <CadIcon />
-              <input
-                ref={refs.password}
-                type={isVisible ? 'text' : 'password'}
-                placeholder="Senha"
-                {...register('password')}
-              />
-              <EyeCloseIcon onClick={() => setIsVisible(prev => !prev)} />
-            </Input>
+            <Input
+              typeField="password"
+              placeholder="Senha"
+              register={register}
+              field="password"
+            />
 
-            <Input onClick={() => refs.confirmPassword.current?.focus()}>
-              <CadIcon />
-              <input
-                ref={refs.confirmPassword}
-                type={confirmIsVisible ? 'text' : 'password'}
-                placeholder="Confirmar senha"
-                {...register('confirmPassword')}
-              />
-              <EyeCloseIcon
-                onClick={() => setConfirmIsVisible(prev => !prev)}
-              />
-            </Input>
+            <Input
+              typeField="confirmPassword"
+              placeholder="Confirmar senha"
+              register={register}
+              field="confirmPassword"
+            />
           </InputContainer>
 
           <ActionButton>
